requirements.txt
opencv-python>=4.8.0
ultralytics>=8.0.0
numpy>=1.24.0
matplotlib>=3.5.0
transformers>=4.30.0
Pillow>=9.0.0
torch>=2.0.0
mediapipe>=0.10.0
